BaSP-TG-133: Created thunk for get tasks
